# executor: try to optimize index merge reader query may send too many cop task