--- a/pkg/executor/test/issuetest/executor_issue_test.go
+++ b/pkg/executor/test/issuetest/executor_issue_test.go
@@ -715,38 +715,42 @@ func TestIssue53221(t *testing.T) {
 	tk.MustExec("drop table if exists t")
 }
 
-func TestIndexLookUpIssue53871(t *testing.T) {
+func TestIndexReaderIssue53871AndIssue54160(t *testing.T) {
 	store := testkit.CreateMockStore(t)
 	tk := testkit.NewTestKit(t, store)
 	tk.MustExec("use test;")
 	tk.MustExec("drop table if exists t;")
-	tk.MustExec("create table t (id int key auto_increment, b int, c int, index idx (b))")
+	tk.MustExec("create table t (id int key auto_increment, b int, c int, index idx (b), index idx2(c))")
 	tk.MustExec(" insert into t () values (), (), (), (), (), (), (), ();")
 	for i := 0; i < 9; i++ {
 		tk.MustExec("insert into t (b) select b from t;")
 	}
-	tk.MustExec(`update t set b = rand() * 10000, c = id;`)
+	tk.MustExec(`update t set b = rand() * 10000, c = rand() * 10000;`)
 	tk.MustQuery("select count(c) from t use index(idx);").Check(testkit.Rows("4096")) // full scan to resolve uncommitted lock.
-	tk.MustQuery("select count(c) from t ignore index(idx)").Check(testkit.Rows("4096"))
+	tk.MustQuery("select count(b) from t use index(idx2);").Check(testkit.Rows("4096"))
+	tk.MustQuery("select count(*) from t ignore index(idx, idx2)").Check(testkit.Rows("4096"))
 	tk.MustExec("analyze table t")
+	// Test Index Lookup Reader query.
 	rows := tk.MustQuery("explain analyze select * from t use index(idx) where b > 0;").Rows()
 	require.Len(t, rows, 3)
 	require.Regexp(t, ".*IndexLookUp.*table_task: {total_time: .*, num: 1, .*", fmt.Sprintf("%v", rows[0]))
 	require.Regexp(t, ".*IndexRangeScan.*rpc_info.*Cop:{num_rpc:1, total_time:.*", fmt.Sprintf("%v", rows[1]))
 	require.Regexp(t, ".*TableRowIDScan.*rpc_info.*Cop:{num_rpc:1, total_time:.*", fmt.Sprintf("%v", rows[2]))
+	// Test Index Merge Reader query.
+	rows = tk.MustQuery("explain analyze select /*+ USE_INDEX_MERGE(t, idx, idx2) */ * from t where b > 5000 or c > 5000;").Rows()
+	require.Len(t, rows, 4)
+	require.Regexp(t, ".*IndexMerge.*table_task:{num:2, .*", fmt.Sprintf("%v", rows[0]))
+	require.Regexp(t, ".*IndexRangeScan.*rpc_info.*Cop:{num_rpc:1, total_time:.*", fmt.Sprintf("%v", rows[1]))
+	require.Regexp(t, ".*IndexRangeScan.*rpc_info.*Cop:{num_rpc:1, total_time:.*", fmt.Sprintf("%v", rows[2]))
+	require.Regexp(t, ".*TableRowIDScan.*rpc_info.*Cop:{num_rpc:2, total_time:.*", fmt.Sprintf("%v", rows[3]))
 }
 
 func TestCalculateBatchSize(t *testing.T) {
-	require.Equal(t, 1024, executor.CalculateBatchSize(true, 5000, 1024, 20000))
-	require.Equal(t, 1024, executor.CalculateBatchSize(true, 1024, 1024, 20000))
-	require.Equal(t, 1024, executor.CalculateBatchSize(true, 10, 1024, 20000))
-	require.Equal(t, 258, executor.CalculateBatchSize(true, 10, 1024, 258))
-	require.Equal(t, 1024, executor.CalculateBatchSize(true, 0, 1024, 20000))
-	require.Equal(t, 20000, executor.CalculateBatchSize(false, 50000, 1024, 20000))
-	require.Equal(t, 20000, executor.CalculateBatchSize(false, 18000, 1024, 20000))
-	require.Equal(t, 8192, executor.CalculateBatchSize(false, 5000, 1024, 20000))
-	require.Equal(t, 1024, executor.CalculateBatchSize(false, 1024, 1024, 20000))
-	require.Equal(t, 1024, executor.CalculateBatchSize(false, 10, 1024, 20000))
-	require.Equal(t, 258, executor.CalculateBatchSize(false, 10, 1024, 258))
-	require.Equal(t, 1024, executor.CalculateBatchSize(false, 0, 1024, 20000))
+	require.Equal(t, 20000, executor.CalculateBatchSize(50000, 1024, 20000))
+	require.Equal(t, 20000, executor.CalculateBatchSize(18000, 1024, 20000))
+	require.Equal(t, 8192, executor.CalculateBatchSize(5000, 1024, 20000))
+	require.Equal(t, 1024, executor.CalculateBatchSize(1024, 1024, 20000))
+	require.Equal(t, 1024, executor.CalculateBatchSize(10, 1024, 20000))
+	require.Equal(t, 258, executor.CalculateBatchSize(10, 1024, 258))
+	require.Equal(t, 1024, executor.CalculateBatchSize(0, 1024, 20000))
 }